+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Version header format: `## [x.x.x] - yyyy-mm-dd`
 - Configuration File template rendering with built-in values
 - Configuration conversions
 - Configuration init with the format option 
+- Semgrep, osemgrep, gitleaks shell commands
+- Code Scan Pipeline
 
 ### Fixed